Move placeholder image and image preview to own files
