+++ site/gatsby-site/src/components/forms/SubmissionWizard/StepTwo.js
@@ -5,9 +5,10 @@ import { Trans, useTranslation } from 'react-i18next';
 import * as yup from 'yup';
 import Label from '../Label';
 import supportedLanguages from '../../i18n/languages.json';
-import { getCloudinaryPublicID, PreviewImageInputGroup } from 'utils/cloudinary';
+import { getCloudinaryPublicID } from 'utils/cloudinary';
 import StepContainer from './StepContainer';
 import TagsInputGroup from '../TagsInputGroup';
+import PreviewImageInputGroup from 'components/forms/PreviewImageInputGroup';
 import FieldContainer from './FieldContainer';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faMedal, faImage, faLanguage } from '@fortawesome/free-solid-svg-icons';
