+++ site/gatsby-site/src/components/submissions/SubmissionForm.js
@@ -4,8 +4,9 @@ import { Spinner } from 'flowbite-react';
 import { useFormikContext } from 'formik';
 import TextInputGroup from '../../components/forms/TextInputGroup';
 import TagsInputGroup from '../../components/forms/TagsInputGroup';
+import PreviewImageInputGroup from '../../components/forms/PreviewImageInputGroup';
 import useToastContext, { SEVERITY } from '../../hooks/useToast';
-import { getCloudinaryPublicID, PreviewImageInputGroup } from '../../utils/cloudinary';
+import { getCloudinaryPublicID } from '../../utils/cloudinary';
 import 'react-bootstrap-typeahead/css/Typeahead.css';
 import Label from '../forms/Label';
 import IncidentIdField from '../../components/incidents/IncidentIdField';
