+++ site/gatsby-site/src/utils/cloudinary.js
@@ -5,11 +5,7 @@ import { format, quality } from '@cloudinary/base/actions/delivery';
 import { auto } from '@cloudinary/base/qualifiers/format';
 import { auto as qAuto } from '@cloudinary/base/qualifiers/quality';
 import config from '../../config';
-import TextInputGroup from '../components/forms/TextInputGroup';
-import { Spinner } from 'flowbite-react';
-import { Trans } from 'react-i18next';
-import hash from 'object-hash';
-import debounce from 'lodash/debounce';
+import PlaceholderImage from 'components/PlaceholderImage';
 
 const getCloudinaryPublicID = (url) => {
   // https://cloudinary.com/documentation/fetch_remote_images#auto_upload_remote_files
@@ -88,226 +84,4 @@ const Image = ({
   }
 };
 
-const PreviewImageInputGroup = ({
-  cloudinary_id,
-  name,
-  label,
-  placeholder,
-  values,
-  errors,
-  touched,
-  handleChange,
-  handleBlur,
-  className = '',
-  schema,
-  alt = '',
-  icon,
-}) => {
-  const [cloudinaryID, setCloudinaryID] = useState(cloudinary_id);
-
-  // Debounced function needs to be in ref
-  // so it can maintain its internal state
-  const debouncedUpdateCloudinaryId = useRef(debounce((c) => setCloudinaryID(c), 2000)).current;
-
-  useEffect(() => {
-    const id = values.cloudinary_id || cloudinary_id;
-
-    if (id && id.length > 0) {
-      // We debounce updating the cloudinary_id to avoid
-      // making too many requests to cloudinary
-      debouncedUpdateCloudinaryId(id);
-    } else {
-      // But if the ID is null, then no request will made,
-      // so we can update right away.
-      setCloudinaryID('');
-    }
-  }, [cloudinary_id, values.cloudinary_id]);
-
-  return (
-    <>
-      <TextInputGroup
-        {...{
-          name,
-          label,
-          icon,
-          placeholder,
-          values,
-          errors,
-          touched,
-          handleChange,
-          className,
-          handleBlur,
-          schema,
-        }}
-      />
-      <figure
-        data-cy="image-preview-figure"
-        id="image-preview-figure"
-        className="text-center md:flex md:items-center md:justify-center md:flex-col"
-      >
-        <div
-          className="grid grid-cols-1 grid-rows-1 items-center justify-center mt-4"
-          style={{ height: '20vh' }}
-        >
-          <div style={{ gridRowStart: '1', gridColumnStart: '1', maxHeight: '100%', zIndex: 1 }}>
-            <Spinner size="xl" />
-          </div>
-          <div
-            style={{
-              gridRowStart: '1',
-              gridColumnStart: '1',
-              height: '20vh',
-              maxWidth: '100%',
-              zIndex: 2,
-            }}
-          >
-            <Image
-              publicID={cloudinaryID}
-              alt={alt}
-              height="300"
-              className="inline-block mx-auto min-h-48 min-w-48 max-w-full h-full bg-white"
-            />
-          </div>
-        </div>
-        <figcaption className="mt-2">
-          <Trans>Selected Image</Trans>
-        </figcaption>
-      </figure>
-    </>
-  );
-};
-
-function PlaceholderImage({ title, siteName, itemIdentifier, height = 480, style, className }) {
-  const canvasRef = useRef();
-
-  const h = Math.floor(Number(height.replace('px', '')));
-
-  const w = Math.floor((h * 4) / 3);
-
-  const random = mulberry32(Number('0x' + hash(itemIdentifier || title)) % Number.MAX_SAFE_INTEGER);
-
-  const randInt = (min, max) => Math.floor(min + random() * (max - min));
-
-  const randomChoice = (list) => list[Math.floor(random() * list.length)];
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-
-    const charHeight = h / 12.5;
-
-    const padding = charHeight * 0.2;
-
-    canvas.width = w;
-    canvas.height = h;
-
-    const ctx = canvas.getContext('2d');
-
-    const colorScheme = randomChoice([
-      { background: '#020241', text: '#ebebff', glow: '#ffffff' },
-      { background: '#030d01', text: '#3bff19', glow: '#ffffff' },
-      { background: '#2d2400', text: '#ffcc00', glow: '#ffffff' },
-      { background: '#fffbd3', text: '#ff5064', glow: '#ffe6f0' },
-    ]);
-
-    // Set background
-    ctx.rect(0, 0, w, h);
-    ctx.fillStyle = colorScheme.background;
-    ctx.fill();
-
-    ctx.font = `${charHeight}px monospace`;
-
-    const charWidth = ctx.measureText('M').width;
-
-    const charsPerLine = Math.floor(w / charWidth);
-
-    const numLines = Math.floor(h / charHeight);
-
-    const title64 = btoa(
-      title
-        ? title.replace(/[^A-Za-z0-9 ]/g, '')
-        : Array(10)
-            .fill()
-            .map(() => Math.floor(random() * Number.MAX_SAFE_INTEGER))
-            .join('')
-    ).replace(/[=+]/g, '');
-
-    let displayText = Array(numLines).fill(title64).join('.');
-
-    // Top half if there's an itemIdentifier, else all
-    const availableLines = itemIdentifier ? numLines / 2 : numLines;
-
-    // Pick randomly from available lines, not counting the first and last
-    const rowIndex = charsPerLine * randInt(1, availableLines - 1);
-
-    // Pick random column such that the full siteName can fit on the line
-    const colIndex = randInt(1, charsPerLine - siteName.length - 1);
-
-    const siteNameIndex = rowIndex + colIndex;
-
-    displayText = insertStringAtIndex(displayText, siteName, siteNameIndex);
-
-    let itemIdentifierIndex;
-
-    if (itemIdentifier) {
-      itemIdentifierIndex =
-        charsPerLine * randInt(numLines / 2, numLines - 3) +
-        randInt(1, charsPerLine - itemIdentifier.length - 1);
-
-      displayText = insertStringAtIndex(displayText, itemIdentifier, itemIdentifierIndex);
-    }
-
-    ctx.filter = `drop-shadow(0px 0px 10px ${colorScheme.glow})`;
-    for (let line = 0; line < numLines; line++) {
-      ctx.fillStyle = setAlpha(colorScheme.text, 0.2 + random() / 4);
-
-      const textIndex = (line * charsPerLine) % displayText.length;
-
-      const lineText = displayText.slice(textIndex, textIndex + charsPerLine);
-
-      const y = (charHeight + padding) * (line + 1);
-
-      ctx.fillText(lineText, padding, y);
-
-      ctx.fillStyle = colorScheme.text;
-      if (line * charsPerLine < siteNameIndex && siteNameIndex < (line + 1) * charsPerLine) {
-        const x = (siteNameIndex % charsPerLine) * charWidth + padding;
-
-        ctx.fillText(siteName, x, y);
-      }
-      if (
-        itemIdentifier &&
-        line * charsPerLine < itemIdentifierIndex &&
-        itemIdentifierIndex < (line + 1) * charsPerLine
-      ) {
-        const x = (itemIdentifierIndex % charsPerLine) * charWidth + padding;
-
-        ctx.fillText(itemIdentifier, x, y);
-      }
-    }
-  }, []);
-
-  return <canvas ref={canvasRef} width={w} height={h} {...{ className, style }}></canvas>;
-}
-
-// Seeded RNG
-function mulberry32(a) {
-  return function () {
-    var t = (a += 0x6d2b79f5);
-
-    t = Math.imul(t ^ (t >>> 15), t | 1);
-    t ^= t + Math.imul(t ^ (t >>> 7), t | 61);
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
-}
-
-function insertStringAtIndex(outerString, innerString, index) {
-  return outerString.slice(0, index) + innerString + outerString.slice(index);
-}
-
-function setAlpha(hexColor, amount) {
-  const alphaHex = Number(Math.floor(amount * 255)).toString(16);
-
-  return hexColor + (alphaHex.length == 1 ? '0' + alphaHex : alphaHex);
-}
-
-export { getCloudinaryPublicID, Image, PreviewImageInputGroup };
+export { getCloudinaryPublicID, Image };
